+++ mobile/android/android-components/docs/changelog.md
@@ -16,6 +16,11 @@ permalink: /changelog/
   * Added `endOfMenuAlwaysVisible` property/parameter to `BrowserMenuBuilder` constructor and to `BrowserMenu.show` function.
     When is set to true makes sure the bottom of the menu is always visible, this allows use cases like [#3211](https://github.com/mozilla-mobile/android-components/issues/3211).
 
+* 🛑 Removed deprecated components (See blog posting):
+  * feature-session-bundling
+  * ui-progress
+  * ui-doorhanger
+
 # 0.56.1
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.56.0...v0.56.1)
 * [Dependencies](https://github.com/mozilla-mobile/android-components/blob/v0.56.1/buildSrc/src/main/java/Dependencies.kt)